% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metainformation.R
\name{.get_metainformation}
\alias{.get_metainformation}
\title{get meta-information from data}
\usage{
.get_metainformation(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{not used}
}
\value{
list
}
\description{
returns list with variable informations for script processing

label: variable-labels

labels: value-labels

na_values: missing-values

na_range: missing-range

scale: scale-level

format: spss-format (e.g. F8.0, A25)
}
