% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metainformation.R
\name{.get_spss_format}
\alias{.get_spss_format}
\title{calculate spss format from data}
\usage{
.get_spss_format(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{not used}
}
\value{
list
}
\description{
calculate spss format from data
}
