% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spss.R
\name{.spss_syntax_missing_range}
\alias{.spss_syntax_missing_range}
\title{prepare syntax: set missing-range}
\usage{
.spss_syntax_missing_range(na_range)
}
\arguments{
\item{na_range}{named list with missing-ranges}
}
\value{
character vector with SPSS syntax
}
\description{
prepare syntax: set missing-range
}
