% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spss.R
\name{.spss_syntax_missing_values}
\alias{.spss_syntax_missing_values}
\title{prepare syntax: set missing-values}
\usage{
.spss_syntax_missing_values(na_values)
}
\arguments{
\item{na_values}{named list with missing-values}
}
\value{
character vector with SPSS syntax
}
\description{
prepare syntax: set missing-values
}
