% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spss.R
\name{.spss_syntax_value_labels}
\alias{.spss_syntax_value_labels}
\title{prepare syntax: set value-labels}
\usage{
.spss_syntax_value_labels(labels)
}
\arguments{
\item{labels}{named list with value-labels}
}
\value{
character vector with SPSS syntax
}
\description{
prepare syntax: set value-labels
}
