% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spss.R
\name{.spss_syntax_variable_labels}
\alias{.spss_syntax_variable_labels}
\title{prepare syntax: set variable-labels}
\usage{
.spss_syntax_variable_labels(label)
}
\arguments{
\item{label}{named list with variable-labels}
}
\value{
character vector with SPSS syntax
}
\description{
prepare syntax: set variable-labels
}
