% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spss.R
\name{.spss_syntax_variable_level}
\alias{.spss_syntax_variable_level}
\title{prepare syntax: set variable-level}
\usage{
.spss_syntax_variable_level(scale)
}
\arguments{
\item{scale}{named list with variable-level}
}
\value{
character vector with SPSS syntax
}
\description{
prepare syntax: set variable-level
}
