% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_char_cols.R
\name{i_fix_char_cols}
\alias{i_fix_char_cols}
\title{merge split string columns into origin column}
\usage{
i_fix_char_cols(x, exclude = NULL, verbose = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{exclude}{character vector with variables not taken into account}

\item{verbose}{print messages to console}
}
\value{
data.frame
}
\description{
Merge split string columns into origin column. By default foreign splits long strings into several columns.
This is a workaround to fix this problem by combining columns split columns into one.
}
\details{
Columns are merged by the following rules:
\enumerate{
\item columns must be character
\item columns must be in sequential order
\item origin column names are extended by 0-9 and A-Z. If a matching variable is found in sequential column order and is character, columns will be merged.
\item if there is a punct character in column name, everything after punct is take into account (e.g. V110_TE will be merged with V110_0 when both are character and in sequential order)
}
}
