% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_foreign.R
\name{i_read_spss}
\alias{i_read_spss}
\title{Read SPSS file}
\usage{
i_read_spss(
  file,
  trim_values = T,
  sort_value_labels = T,
  return_data_frame = T,
  ...
)
}
\arguments{
\item{file}{file path}

\item{trim_values}{trim trailing spaces from value labels}

\item{sort_value_labels}{sort value labels}

\item{return_data_frame}{return data as data.frame or list}

\item{...}{arguments passed to foreign::read.spss}
}
\value{
data.frame
}
\description{
Read SPSS data files and add class i_labelled.
Wrapper for foreign::read.spss.
}
