% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spss.R
\name{i_write_spss}
\alias{i_write_spss}
\title{write SPSS syntax}
\usage{
i_write_spss(
  x,
  data,
  syntax,
  dec = c(".", ","),
  delimiter = c("\\t", ",", ";", "~"),
  qualifier = c("double", "escape"),
  encoding = "UTF-8",
  add_syntax_top = NULL,
  add_syntax_bottom = NULL,
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{data}{file path for data file}

\item{syntax}{file path for syntax file}

\item{dec}{decimal point}

\item{delimiter}{delimiter used in data file}

\item{qualifier}{how to deal with embedded double quotes}

\item{encoding}{file encoding}

\item{add_syntax_top}{add SPSS syntax before the default syntax}

\item{add_syntax_bottom}{add SPSS syntax after the default syntax}

\item{...}{not used}
}
\value{
no return value. writes data and spss syntax to file.
}
\description{
write SPSS syntax
}
