% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output.R
\name{.get_metainformation}
\alias{.get_metainformation}
\title{get meta-information from data}
\usage{
.get_metainformation(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{not used}
}
\value{
list
}
\description{
get meta-information from data
}
