% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_foreign.R
\name{.read_foreign}
\alias{.read_foreign}
\title{Read data via R foreign}
\usage{
.read_foreign(
  file,
  trim_values = T,
  to.data.frame = F,
  use.value.labels = F,
  ...
)
}
\arguments{
\item{file}{file path}

\item{trim_values}{trim trailing spaces from value labels}

\item{to.data.frame}{convert list to data.frame}

\item{use.value.labels}{convert to factor when data has value labels}

\item{...}{arguments passed to foreign::read.spss}
}
\value{
List with Data
}
\description{
Wrapper for foreign::read.spss.
Read SPSS Daten files before adding class i_labelled.
}
