% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spss.R
\name{i_read_spss}
\alias{i_read_spss}
\title{Read SPSS file}
\usage{
i_read_spss(
  file,
  trim_values = TRUE,
  sort_value_labels = TRUE,
  fix_duplicate_labels = TRUE,
  return_data_frame = TRUE,
  warn = TRUE,
  ...
)
}
\arguments{
\item{file}{file path}

\item{trim_values}{trim trailing spaces from value labels}

\item{sort_value_labels}{sort value labels}

\item{fix_duplicate_labels}{duplicate value labels will be fixed: replace "" with value; replace duplicate labels with label + '\emph{duplicated}' + value}

\item{return_data_frame}{return data as data.frame or list}

\item{warn}{show warnings}

\item{...}{arguments passed to foreign::read.spss}
}
\value{
data as list or data.frame
}
\description{
read SPSS data files and add class i_labelled.

wrapper for foreign::read.spss.
}
