% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_spss_syntax.R
\name{i_write_spss}
\alias{i_write_spss}
\title{write SPSS syntax}
\usage{
i_write_spss(
  x,
  data,
  syntax,
  delimiter = "\\t",
  qualifier = c("double", "excape"),
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{data}{file path for data file}

\item{syntax}{file path for syntax file}

\item{delimiter}{delimiter used in data file}

\item{qualifier}{how to deal with embedded double quotes}

\item{encoding}{file encoding}

\item{...}{not used}
}
\value{
something
}
\description{
write SPSS syntax
}
